pyproject.toml
[project]
name = "chest_xray_classifier"
version = "0.1.0"
description = "Production-grade 3-class chest X-ray classifier (normal / bacterial pneumonia / viral pneumonia)."
requires-python = ">=3.13"
authors = [
  { name = "Daniil Kiselyov", email = "your-email@example.com" },
]
license = { text = "MIT" }
readme = "README.md"

dependencies = [
  "hydra-core>=1.3.2",
  "omegaconf>=2.3.0",
  "pyyaml>=6.0.1",
  "structlog>=24.1.0",
  "rich>=13.7.0",
"torch>=2.2.0",
  "torchvision>=0.17.0",
  "lightning>=2.2.0",
  "torchmetrics>=1.3.0",
]

[dependency-groups]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=4.1.0",
  "ruff>=0.3.0",
  "mypy>=1.9.0",
  "pre-commit>=3.6.0",
  "nbqa>=1.8.0",
  "nbstripout>=0.7.0",
  "ipykernel>=6.29.0",
  "matplotlib>=3.8.0",
]
serving = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.29.0",
  "pydantic>=2.6.0",
  "python-multipart>=0.0.9",
  "prometheus-fastapi-instrumentator>=7.0.0",
  "huggingface-hub>=0.22.0",
"pillow>=10.2.0",
]
tracking = [
  "mlflow>=2.11.0",
  "dvc>=3.48.0",
  "wandb>=0.16.0",
]
docs = [
  "mkdocs-material>=9.5.0",
  "mkdocstrings[python]>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chest_xray_classifier"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM", "RUF", "N", "D"]
ignore = ["D100", "D104", "D107"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short --cov=src --cov-report=term --cov-report=xml --cov-fail-under=10"

[tool.coverage.run]
source = ["src/chest_xray_classifier"]
branch = true

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "if TYPE_CHECKING:", "raise NotImplementedError"]
